Set conditions for ui update & showEditBtn
1) Update Order & Print Button
Set conditions in the ui whereby only shows when those conditions are met.
2) showEditBtn
Set conditions when select any tables, however use useEffect to set it false when certain conditions are met. This is due to the slow update after selected table. And to avoid the showEditBtn when added new items during updating.